FootballScoreKeeper
This is my Scorekeeper app for tracking scores in the football game. It is shared for free for non-commercial usage.

Features include:

multiple amounts of points that can be scored;
app is divided into two columns, one for each team;
each column contains a large TextView to keep track of the current score for that team.
each column contains multiple buttons to track goal, yellow and red cards;
the layout contains a ‘reset’ button;
the abiliity to type team names (EditText feature);
scroll view.
The code adheres to all of the following best practices:

Text sizes are defined in sp; Lengths are defined in dp; Padding and margin is used appropriately, such that the views are not crammed up against each other; All methods are named by their intended effect or in the style required by a callback interface; There are no unnecessary blank lines.One variable is declared per declaration line.

For additional information, or if you wish to cooperate, please contact directly!